requirements.txt
certifi==2016.2.28
falcon
future
gunicorn
peewee
python-mimeparse
python-telegram-bot
six
urllib3
wheel
